Update integration test
