+++ odw/core/etl/transformation/curated/appeal_attribute_matrix_curated_process.py
@@ -3,6 +3,8 @@ from typing import Any
 
 
 class AppealAttributeMatrixCuratedProcess(CurationProcess):
+    STANDARDISED_TABLE = "appeal_attribute_matrix"
+    HARMONISED_TABLE = "ref_appeal_attribute_matrix"
     OUTPUT_TABLE = "ref_appeal_attribute_matrix"
 
     def __init__(self, spark):
